Add different cell and background colors
When a new level starts up, it just picks these colors at random
but you can press c to cycle forward and x to cycle backwards
and a prints out Approved to the console. The idea is that I
will run through permutations and pick out the ones I like, then
parse the log later to make the array of approved colors.